requirements.txt
streamlit
python-dotenv
langchain
langchain-community
huggingface_hub
transformers
torch
faiss-cpu
PyPDF2
